refactor transmit channel
