+++ src/main.rs
@@ -1,8 +1,7 @@
 use color_eyre::eyre::Context;
 use eid_server::{
-    config::Config,
+    config::{Config, TransmitConfig},
     domain::eid::service::{EIDServiceConfig, UseidService},
-    sal::transmit::config::TransmitConfig,
     server::{Server, ServerConfig},
     telemetry,
 };
