+++ src/sal/transmit/mod.rs
@@ -1,5 +1,4 @@
 pub mod channel;
-pub mod config;
 pub mod error;
 pub mod protocol;
 pub mod session;
